cdvrp working for multiple routes
